+++ .github/workflows/test-pr.yml
@@ -71,7 +71,7 @@ jobs:
     - name: Run Controllers tests
       run: make test-controllers
 
-  api-unit-tests:
+  api-tests:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
@@ -90,25 +90,4 @@ jobs:
         go-version: '^1.17'
 
     - name: Run API unit tests
-      run: make test-api-unit
-
-  api-integration-tests:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v3
-
-    - uses: actions/cache@v3
-      with:
-        path: |
-          ~/.cache/go-build
-          ~/go/pkg/mod
-        key: ${{ runner.os }}-go-${{ hashFiles('api/go.sum') }}
-        restore-keys: |
-          ${{ runner.os }}-go-
-
-    - uses: actions/setup-go@v3
-      with:
-        go-version: '^1.17'
-
-    - name: Run API integration tests
-      run: make test-api-integration
+      run: make test-api
