+++ Makefile
@@ -76,13 +76,8 @@ test-unit: test-controllers test-api-unit
 test-controllers: install-ginkgo manifests-controllers generate-controllers fmt vet ## Run tests.
 	cd controllers && GINKGO_NODES=$(CONTROLLERS_GINKGO_NODES) ../scripts/run-tests.sh
 
-test-api: test-api-unit test-api-integration
-
-test-api-unit: install-ginkgo fmt vet
+test-api: install-ginkgo fmt vet
 	cd api && ../scripts/run-tests.sh --skip-package=test
-
-test-api-integration: install-ginkgo
-	cd api && ../scripts/run-tests.sh tests/integration
 
 test-e2e: install-ginkgo
 	cd api && ../scripts/run-tests.sh tests/e2e
