+++ api/authorization/retrying_k8s_client.go
@@ -1,4 +1,4 @@
-package repositories
+package authorization
 
 import (
 	"context"
