+++ api/authorization/testhelpers/auth_provider.go
@@ -1,4 +1,4 @@
-package helpers
+package testhelpers
 
 import (
 	"crypto/rand"
