+++ api/repositories/process_repository.go
@@ -20,7 +20,7 @@ const (
 	processPrefix       = "cf-proc-"
 )
 
-func NewProcessRepo(namespaceRetriever NamespaceRetriever, userClientFactory UserK8sClientFactory, namespacePermissions *authorization.NamespacePermissions) *ProcessRepo {
+func NewProcessRepo(namespaceRetriever NamespaceRetriever, userClientFactory authorization.UserK8sClientFactory, namespacePermissions *authorization.NamespacePermissions) *ProcessRepo {
 	return &ProcessRepo{
 		namespaceRetriever:   namespaceRetriever,
 		clientFactory:        userClientFactory,
@@ -30,7 +30,7 @@ func NewProcessRepo(namespaceRetriever NamespaceRetriever, userClientFactory Use
 
 type ProcessRepo struct {
 	namespaceRetriever   NamespaceRetriever
-	clientFactory        UserK8sClientFactory
+	clientFactory        authorization.UserK8sClientFactory
 	namespacePermissions *authorization.NamespacePermissions
 }
