+++ api/tests/e2e/whoami_test.go
@@ -5,7 +5,7 @@ import (
 	"encoding/base64"
 	"net/http"
 
-	"code.cloudfoundry.org/korifi/api/tests/helpers"
+	"code.cloudfoundry.org/korifi/tests/helpers"
 	"github.com/go-resty/resty/v2"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
