+++ tests/helpers/cert_auth_header.go
